Add ability to add/remove TLVs via UI

+++ at.innovative-solutions.tlvVisualizer/src/at/innovative_solutions/tlv/ID.java
@@ -3,45 +3,44 @@ package at.innovative_solutions.tlv;
 import java.nio.BufferUnderflowException;
 import java.nio.ByteBuffer;
 
-// TODO implement equals, simplify test cases
 public class ID {
 	public static final int CLASS_UNIVERSAL = 0;
 	public static final int CLASS_APPLICATION = 1;
 	public static final int CLASS_CONTEXT = 2;
 	public static final int CLASS_PRIVATE = 3;
-	
+
 	final int _tagClass;
 	final boolean _isPrimitive;
 	final int _tagNumber;
 	final int _longFormBytes;
-	
+
 	public ID(int tagClass, boolean isPrimitive, int tagNumber) {
 		this(tagClass, isPrimitive, tagNumber, 0);
 	}
-	
+
 	public ID(int tagClass, boolean isPrimitive, int tagNumber, int longFormBytes) {
 		_tagClass = tagClass;
 		_isPrimitive = isPrimitive;
 		_tagNumber = tagNumber;
 		_longFormBytes = longFormBytes;
 	}
-	
+
 	public int getTagClass() {
 		return _tagClass;
 	}
-	
+
 	public boolean isPrimitive() {
 		return _isPrimitive;
 	}
-	
+
 	public int getTagNumber() {
 		return _tagNumber;
 	}
-	
+
 	public int getLongFormByteCnt() {
 		return _longFormBytes;
 	}
-	
+
 	public byte[] toBytes() {
 		if(_tagNumber <= 30 && _longFormBytes == 0) {
 			int b = _tagClass << 6 | (_isPrimitive ? 0 : 1 << 5) | _tagNumber;
@@ -57,11 +56,11 @@ public class ID {
 				final byte blockData = (byte) ((_tagNumber >> ((nrOfBlocks - i) * 7)) & 0x7F);
 				result[i] = (byte) ((i != nrOfBlocks ? (byte)0x80 : (byte)0) | blockData);
 			}
-			
+
 			return result;
 		}
 	}
-	
+
 	// TODO correct exception type
 	public long toLong() {
 		byte[] bytes = toBytes();
@@ -72,16 +71,16 @@ public class ID {
 			result = result << 8 | (bytes[n] & 0xff);
 		return result;
 	}
-	
+
 	// TODO protect against overflow
 	public static ID parseID(ByteBuffer octets) {
 		try {
-			byte first = (byte) octets.get();
-			
+			byte first = octets.get();
+
 			final int tagClass = (first >> 6) & 0x3;
 			final boolean isPrimitive = (first & (1 << 5)) == 0;
 			int tagNumber = first & 0x1F;
-			
+
 			int longFormBytes = 0;
 			if(tagNumber == 0x1F) {
 				tagNumber = 0;
@@ -92,32 +91,32 @@ public class ID {
 					longFormBytes++;
 				} while((piece & 0x80) != 0);
 			}
-			
+
 			return new ID(tagClass, isPrimitive, tagNumber, longFormBytes);
 		} catch(BufferUnderflowException e) {
 			throw new ParseError("id", "Not enough bytes to extract ID", e);
 		}
 	}
-	
+
 	public boolean equals(final Object other) {
 		if(!(other instanceof ID))
 			return false;
-		
+
 		return this._tagClass == ((ID)other)._tagClass
 				&& this._isPrimitive == ((ID)other)._isPrimitive
 				&& this._tagNumber == ((ID)other)._tagNumber
 				&& this._longFormBytes == ((ID)other)._longFormBytes;
 	}
-	
+
 	public boolean equalContents(final Object other) {
 		if(!(other instanceof ID))
 			return false;
-	
+
 		return this._tagClass == ((ID)other)._tagClass
 				&& this._isPrimitive == ((ID)other)._isPrimitive
 				&& this._tagNumber == ((ID)other)._tagNumber;
 	}
-	
+
 	public String toString() {
 		StringBuffer str = new StringBuffer();
 		str.append("ID(");
@@ -127,7 +126,7 @@ public class ID {
 		str.append(", ");
 		str.append(_tagNumber);
 		str.append(")");
-		
+
 		return str.toString();
 	}
 }
